requirements.txt
asgiref==3.8.1
Django==5.0.4
django-redis==5.4.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0
redis==5.0.3
sqlparse==0.5.0
tzdata==2024.1
